# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cutie Task Management API
  description: |
    Cutie后端API文档 - 完整的任务管理系统API
    
    ## 概述
    
    Cutie API提供了完整的任务管理功能，包括：
    - 任务的CRUD操作
    - 日程安排和管理
    - 时间块管理
    - 模板系统
    - 领域层级管理
    - 排序和组织
    
    ## 架构特点
    
    - **RESTful设计**: 遵循REST API最佳实践
    - **类型安全**: 严格的数据类型定义
    - **统一错误处理**: 一致的错误响应格式
    - **分页支持**: 大数据集的分页查询
    - **实时更新**: 支持WebSocket推送（未来版本）
    
    ## 认证
    
    V1.0版本为单用户桌面应用，暂不需要认证。未来版本将支持JWT认证。
    
    ## 错误处理
    
    所有错误响应都遵循统一格式，包含错误类型、消息、详细信息和错误代码。
    
    ## 数据格式
    
    - 所有时间戳使用ISO 8601格式 (RFC 3339)
    - 所有ID使用UUID v4格式
    - JSON字段用于复杂数据结构
    
  version: 1.0.0
  contact:
    name: Cutie Development Team
    email: dev@cutie.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3030/api
    description: 开发环境服务器
  - url: http://127.0.0.1:3030/api
    description: 本地环境服务器
paths:
...
  /time-blocks:
    post:
      tags:
        - TimeBlocks
      summary: 创建时间块
      description: |
        创建一个新的时间块，并可选择性地链接任务。
        
        **业务逻辑**:
        - 验证时间范围的有效性
        - 检查时间冲突
        - 自动链接指定的任务
      operationId: createTimeBlock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeBlockPayload'
      responses:
        '201':
          description: 时间块创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TimeBlock'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags:
        - TimeBlocks
      summary: 获取时间块列表
      description: |
        根据不同条件获取时间块列表。
        
        **查询模式**:
        - date: 获取特定日期的时间块
        - start_date + end_date: 获取日期范围内的时间块
        - task_id: 获取与任务关联的时间块
      operationId: getTimeBlocks
      parameters:
        - name: date
          in: query
          description: 特定日期
          required: false
          schema:
            type: string
            format: date-time
        - name: start_date
          in: query
          description: 开始日期
          required: false
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: 结束日期
          required: false
          schema:
            type: string
            format: date-time
        - name: task_id
          in: query
          description: 任务ID
          required: false
          schema:
            type: string
            format: uuid
        - name: area_id
          in: query
          description: 领域ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 时间块列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimeBlock'
        '422':
          $ref: '#/components/responses/ValidationError'
  /time-blocks/{id}:
    get:
      tags:
        - TimeBlocks
      summary: 获取时间块详情
      operationId: getTimeBlock
      parameters:
        - name: id
          in: path
          required: true
          description: 时间块ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 时间块详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TimeBlock'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - TimeBlocks
      summary: 更新时间块
      operationId: updateTimeBlock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTimeBlockPayload'
      responses:
        '200':
          description: 时间块更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TimeBlock'

    delete:
      tags:
        - TimeBlocks
      summary: 删除时间块
      operationId: deleteTimeBlock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 时间块删除成功
  /time-blocks/{id}/tasks:
    post:
      tags:
        - TimeBlocks
      summary: 链接任务到时间块
      operationId: linkTaskToBlock
      parameters:
        - name: id
          in: path
          required: true
          description: 时间块ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task_id:
                  type: string
                  format: uuid
              required:
                - task_id
      responses:
        '204':
          description: 任务链接成功
  /time-blocks/{id}/tasks/{taskId}:
    delete:
      tags:
        - TimeBlocks
      summary: 取消任务关联
      operationId: unlinkTaskFromBlock
      parameters:
        - name: id
          in: path
          required: true
          description: 时间块ID
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          description: 任务ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 任务关联取消成功
  /time-blocks/conflicts:
    get:
      tags:
        - TimeBlocks
      summary: 检查时间冲突
      description: 检查指定时间范围是否与现有时间块冲突
      operationId: checkTimeConflict
      parameters:
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: exclude_id
          in: query
          required: false
          description: 排除的时间块ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 冲突检查结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      has_conflict:
                        type: boolean
                      start_time:
                        type: string
                        format: date-time
                      end_time:
                        type: string
                        format: date-time
  /time-blocks/free-slots:
    get:
      tags:
        - TimeBlocks
      summary: 查找空闲时间段
      description: 在指定时间范围内查找未被时间块占用的空闲时间段
      operationId: findFreeSlots
      parameters:
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: min_duration_minutes
          in: query
          required: true
          description: 最小时长（分钟）
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: 空闲时间段列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FreeTimeSlot'

  # ==================== 模板管理 API ====================
...
components:
...
  schemas:
...
    TimeBlock:
      type: object
      description: 时间块实体
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        glance_note:
          type: string
          nullable: true
        detail_note:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        area_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean
      required:
        - id
        - start_time
        - end_time
        - created_at
        - updated_at
        - is_deleted
...
    FreeTimeSlot:
      type: object
      description: 空闲时间段
      properties:
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          description: 时长（分钟）
      required:
        - start_time
        - end_time
        - duration_minutes

    # ==================== 枚举类型 ====================
...
    CreateTimeBlockPayload:
      type: object
      description: 创建时间块请求载荷
      properties:
        title:
          type: string
          nullable: true
        glance_note:
          type: string
          nullable: true
        detail_note:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        area_id:
          type: string
          format: uuid
          nullable: true
        task_ids:
          type: array
          description: 要关联的任务ID列表
          items:
            type: string
            format: uuid
      required:
        - start_time
        - end_time
        - task_ids
    UpdateTimeBlockPayload:
      type: object
      description: 更新时间块请求载荷
      properties:
        title:
          type: string
          nullable: true
        glance_note:
          type: string
          nullable: true
        detail_note:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
          nullable: true
        end_time:
          type: string
          format: date-time
          nullable: true
        area_id:
          type: string
          format: uuid
          nullable: true
...
    ErrorResponse:
      type: object
      description: 错误响应结构
      properties:
        error_type:
          type: string
          description: 错误类型
        message:
          type: string
          description: 错误消息
        details:
          type: object
          nullable: true
          description: 详细信息
        code:
          type: string
          nullable: true
          description: 错误代码
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          nullable: true
      required:
        - error_type
        - message
        - timestamp
...
  responses:
    ValidationError:
      description: 输入数据验证失败
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error_type:
                    type: string
                    enum: [ValidationError]
                  code:
                    type: string
                    enum: [VALIDATION_FAILED]
          example:
            error_type: "ValidationError"
            message: "输入数据验证失败"
            details:
              validation_errors:
                - field: "title"
                  message: "Title cannot be empty"
                  code: "TITLE_EMPTY"
            code: "VALIDATION_FAILED"
            timestamp: "2024-09-29T10:00:00Z"
    NotFound:
      description: 请求的资源未找到
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error_type:
                    type: string
                    enum: [NotFound]
                  code:
                    type: string
                    enum: [NOT_FOUND]
          example:
            error_type: "NotFound"
            message: "Task with id 123e4567-e89b-12d3-a456-426614174000 not found"
            details:
              entity_type: "Task"
              entity_id: "123e4567-e89b-12d3-a456-426614174000"
            code: "NOT_FOUND"
            timestamp: "2024-09-29T10:00:00Z"
...